# api/swagger/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для маркетплейса на Go с JWT аутентификацией, управлением магазинами, продуктами и категориями",
        "title": "Go Shop API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Go Shop API Support",
            "url": "https://github.com/ReversIncorp/go-shop",
            "email": "a.savko.developer@goshop.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/stores": {
            "post": {
                "description": "Создает новый магазин для пользователя",
                "tags": [
                    "stores"
                ],
                "summary": "Создание магазина",
                "parameters": [
                    {
                        "description": "Данные магазина",
                        "name": "store",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.Store"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Магазин создан",
                        "schema": {
                            "$ref": "#/definitions/entities.Store"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    }
                }
            }
        },
        "/stores/search": {
            "post": {
                "description": "Возвращает список магазинов с пагинацией и фильтрацией",
                "tags": [
                    "stores"
                ],
                "summary": "Получение магазинов по фильтрам",
                "parameters": [
                    {
                        "description": "Параметры поиска",
                        "name": "searchParams",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.StoreSearchParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список магазинов",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    }
                }
            }
        },
        "/stores/{store_id}": {
            "get": {
                "description": "Возвращает информацию о магазине по его ID",
                "tags": [
                    "stores"
                ],
                "summary": "Получение магазина по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID магазина",
                        "name": "store_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о магазине",
                        "schema": {
                            "$ref": "#/definitions/entities.Store"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Магазин не найден",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет информацию о магазине",
                "tags": [
                    "stores"
                ],
                "summary": "Обновление магазина",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID магазина",
                        "name": "store_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновленные данные магазина",
                        "name": "store",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.Store"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Магазин обновлен",
                        "schema": {
                            "$ref": "#/definitions/entities.Store"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Магазин не найден",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет магазин из системы",
                "tags": [
                    "stores"
                ],
                "summary": "Удаление магазина",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID магазина",
                        "name": "store_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Магазин удален",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Магазин не найден",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    }
                }
            }
        },
        "/stores/{store_id}/categories": {
            "get": {
                "description": "Возвращает все категории конкретного магазина",
                "tags": [
                    "categories"
                ],
                "summary": "Получение категорий магазина",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID магазина",
                        "name": "store_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список категорий",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Category"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный ID магазина",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "description": "Связывает категорию с магазином",
                "tags": [
                    "stores"
                ],
                "summary": "Привязка категории к магазину",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID магазина",
                        "name": "store_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ID категории",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Категория привязана",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    }
                }
            }
        },
        "/stores/{store_id}/categories/{category_id}": {
            "delete": {
                "description": "Отвязывает категорию от магазина",
                "tags": [
                    "stores"
                ],
                "summary": "Отвязка категории от магазина",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID магазина",
                        "name": "store_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID категории",
                        "name": "category_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Категория отвязана",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверные ID",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/errorhandling.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entities.Category": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "entities.Store": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entities.StoreSearchParams": {
            "type": "object",
            "required": [
                "cursor",
                "limit"
            ],
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "cursor": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "errorhandling.ResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP статус-код",
                    "type": "integer"
                },
                "details": {
                    "description": "Сообщение для пользователя",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите \"Bearer\" за которым следует пробел и JWT токен.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}